Some tweaks from rebuilding my config
* Increment the default font size.
* Fix dired-preview hook.
* Trying Melpa unstable for now, some packages weren't available when
  I rebuilt.
* Only load work stuff if it exists.
* Delete Clojure stuff.

The documentation/maual for dired-preview doesn't seem to have this
referenced with installation via `use-package`, but I found this in
the author's config:
https://protesilaos.com/emacs/dotemacs#h:c345ff72-d905-482a-a885-fc111252aacf

I'm not using Clojure for anything yet, and our team hasn't settled on
a particular development pattern as a whole anyway. Once we do I'll
re-add Clojure configuration and setup as we develop that
together. The author of Clojure for the Brave and True has a
comprehensive configuration, but it's entirely based on SetupEl, which
I recently tried and removed from here.